Perform offline path replacement to `tokenizer`
If `tokenizer` is a Hugging Face model, vLLM attempts to access Hugging
Face even if the tokenizer is already available offline.  It prevents
specifying a Hugging Face model as `tokenizer` on the offline mode
(i.e. `HF_HUB_OFFLINE` is true).

With this commit, vLLM performs offline path replacement also on
`tokenizer`, not only on `model`.

A test case is added because error related to the offline mode only
occurs when the model and the tokenizer are different.

Signed-off-by: Tsukasa OI <floss_llm@irq.a4lg.com>