Fix LLMEngine.del dp_group cleanup condition
`LLMEngine.__del__` previously used the walrus operator together with
`and`, which caused `dp_group` to be assigned the result of the boolean
expression instead of the actual DP group object due to operator precedence.
As a result, `stateless_destroy_torch_distributed_process_group` could
receive `True`/`False` instead of a process group instance.

Signed-off-by: Yongtao Huang <yongtaoh2022@gmail.com>