Add TODO for future GGUF + MoE handling
GGUF may have unquantized weights (e.g. BF16) and that may be helpful to
separate quantization-specific errors from GGUF-specific errors.

Still, there's no support for MoE layers and checking whether quantization
is required is not easy with the current architecture.

So, this commit just adds a TODO comment to denote that there's
something not implemented yet (for clarity).

Signed-off-by: Tsukasa OI <floss_llm@irq.a4lg.com>