[Bugfix] Use PIECEWISE cudagraph on gpt-oss Ampere
Multiple user reports have come in about repeated, infinite generations
by gpt-oss models on Ampere hardware. After reproducing this on Ampere
hardware, I was able to workaround it by downgrading to PIECEWISE
cudagraphs. FULL and FULL_AND_PIECEWISE can reliably reproduce infinite
generation on the 2nd and later requests.

The check here to downgrade only triggers for gpt-oss models on Ampere
or older CUDA GPUs. This matches where I was able to reproduce the issue
and with what appears to be the largest number of user reports.

There could potentially be problems on more MoE models or with a
slightly larger range of GPUs, but thought it best to make the most
targeted fix for now just to get these models working reliably again on
Ampere GPUs.

Users report that gpt-oss models on vLLM 0.10.2 worked well on Ampere
GPUs, which is further evidence that this is related to us moving to
FULL_AND_PIECEWISE cudagraphs by default with vLLM v0.11.0 that
negatively impacted this hardware.

Signed-off-by: Ben Browning <bbrownin@redhat.com>