[Attention][AMD] Make flash-attn optional
flash-attn is not supported on gfx11. (`VLLM_ATTENTION_BACKEND=FLASH_ATTN` gives `Attention backend FLASH_ATTN is not supported on ROCm`). But if flash-attn is not installed, vllm would fail with "Rocm platform requires upstream flash-attn to be installed" due to the unconditional import in eagle.py.
This PR delays the error to the first use to allow
uses of vLLM without flash-attn to succeed.

Signed-off-by: Matthias Gehre <matthias.gehre@amd.com>