[ROCm] Dynamically Quantize Quark MXFP4 weights at load time
Normally Quark models will quantize most if not all of the weights statically
and store them in the quark model.  In some cases they will exclude certain layers
and leave them unquantized.  This PR will allow us to quantize the weights on the fly
before executing the model.  This is primarily for Deepseek FP4 models where the projection
layers are not statically quantized

Signed-off-by: Douglas Lehr <Doug.Lehr@amd.com>