Issue with custom_all_reduce and cudagraph_mode FULL
Full CUDA graph capture + replay (the default since vLLM v0.11.0) could
lead to memory correctness issues in the custom_all_reduce
implementation due to previous logic that skipped running the actual
all_reduce operation during the compile and warmup phase. This resulted
in compile and warmup seeing a memory allocation, but not observing the
custom op call or the actual reduce operation which could lead to
optmizations being taken assuming purely functional code with no
side-effects when in fact the reduce operation inherently has side
effects.

To fix this, instead of allocating an empty tensor to mimic the reduce
operation, we now call the actual all_reduce during compile and warmup
as well.

Multiple users reported an issue that tracks back to this root cause on
Ampere hardware, which is one platform where we use custom_all_reduce by
default if tensor parallelism is used and NVLink is available.

Examples:
- https://github.com/vllm-project/vllm/issues/26480
- https://github.com/vllm-project/vllm/issues/29998
- https://github.com/vllm-project/vllm/issues/30498

Note that until I rebased this on top of a very recent main, I also had
to tag the all_reduce custom op registration in parallel_state.py with
`torch.Tag.maybe_aliasing_or_mutating` to fix the reproducer described
in #29998. The tag is no longer required on latest main, although I have
no identified exactly why or what changed that fixed that.

Signed-off-by: Ben Browning <bbrownin@redhat.com>