OffloadingConnector: Support kernel_block_size != block_size
This commit enables the offloading connector to work
for cases where the kernel block size is different than vLLM's
logical block size.
To efficiently copy blocks when the logical block size is greater
than the block size, we add an additional parameter to the swap_blocks
function that determines the size per each contiguous copy.

Signed-off-by: Or Ozeri <oro@il.ibm.com>