[BugFix] Handle errors when preprocessing added requests
preprocess_add_request() runs in the engine core input socket processing thread. It's not expected to raise exceptions but if it does, they aren't caught or logged - the input processing thread exits and the engine hangs silently :(

This PR catches and logs request-scoped preprocessing errors, and returns an output to fail the request in question.

I'll open another PR to deal with any other unexpected exceptions occurring in the input socket processing thread (which should be considered fatal, but we should log and shutdown rather than hanging silently!).

Signed-off-by: Nick Hill <nhill@redhat.com>